added another dialog
