+++ kivy_based/Gui.kv
@@ -35,6 +35,7 @@
         size_hint: .8, .2
         orientation: 'vertical'
         TextInput:
+            id: StreamTitleInput
             hint_text: "Stream Title"
             font_size: 30
         BoxLayout:
@@ -45,11 +46,12 @@
                 size_hint: .3, 1
                 on_release: StartupControl.parent.parent.current = "MainScreen"
             Button:
+                id: SubmitButton
                 size_hint: .7, 1
                 pos_hint: {"right": 1}
-                id: "SubmitButton"
                 text: "Setup Stream"
                 on_release: StartupControl.parent.parent.current = "MainScreen"
+                on_press: root.on_submit(StreamTitleInput.text)
 
 <StreamController>:
     canvas:
