+++ kivy_based/exceptions.py
@@ -1,0 +1,8 @@
+class PopupError(Exception):
+    pass
+
+class PopupNotExist(PopupError):
+    pass
+
+class PopupClosed(PopupError):
+    pass
